refactor: retire AppContext in favor of hooks and stores
