+++ src/renderer/__helpers__/test-utils.tsx
@@ -1,25 +1,26 @@
 import { render } from '@testing-library/react';
-import { type ReactElement, type ReactNode, useMemo } from 'react';
+import type { ReactElement } from 'react';
 import { type InitialEntry, MemoryRouter } from 'react-router-dom';
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
-import { AppContext, type AppContextState } from '../context/context';
+import { useShortcutRegistrationStore } from '../hooks/useShortcutRegistration';
 import { type FiltersStore, useAccountsStore, useFiltersStore, useSettingsStore } from '../stores';
 
 import type { Account, SettingsState } from '../types';
 
+import {
+  type LoginsState,
+  type NotificationsState,
+  setIsOnlineOverride,
+  setLoginsOverrides,
+  setNotificationsOverrides,
+} from './hook-mocks';
+
 export { navigateMock } from './vitest.setup';
 export type DeepPartial<T> = { [K in keyof T]?: DeepPartial<T[K]> };
 
-const EMPTY_APP_CONTEXT: TestAppContext = {};
-
-/**
- * Test context
- */
-type TestAppContext = Partial<AppContextState>;
-
-interface RenderOptions extends TestAppContext {
+interface RenderOptions extends Partial<NotificationsState>, Partial<LoginsState> {
   /** Supports path strings or location objects (e.g. with `state` for re-auth). */
   initialEntries?: InitialEntry[];
   /** Seed the accounts store with these accounts. */
@@ -28,81 +29,46 @@ interface RenderOptions extends TestAppContext {
   settings?: Partial<SettingsState>;
   /** Seed the filters store. */
   filters?: Partial<FiltersStore>;
+  /** Override the mocked `useNotifications().refetchNotifications`. */
+  fetchNotifications?: NotificationsState['refetchNotifications'];
+  /** Override the mocked online status. */
+  isOnline?: boolean;
+  /** Seed the shortcut registration error state. */
+  shortcutRegistrationError?: string | null;
 }
 
-/**
- * Props for the AppContextProvider wrapper
- */
-interface AppContextProviderProps {
-  readonly children: ReactNode;
-  readonly value?: TestAppContext;
-  readonly initialEntries?: InitialEntry[];
-}
-
-/**
- * Wrapper component that provides MemoryRouter and AppContext
- * with sensible defaults for testing.
- */
-function AppContextProvider({
-  children,
-  value = EMPTY_APP_CONTEXT,
-  initialEntries,
-}: AppContextProviderProps) {
-  const defaultValue: TestAppContext = useMemo(() => {
-    return {
-      notifications: [],
-      notificationCount: 0,
-      unreadNotificationCount: 0,
-      hasNotifications: false,
-      hasUnreadNotifications: false,
-
-      status: 'success',
-      globalError: undefined,
-
-      // Default mock implementations for all required methods
-      loginWithDeviceFlowStart: vi.fn(),
-      loginWithDeviceFlowPoll: vi.fn(),
-      loginWithDeviceFlowComplete: vi.fn(),
-      loginWithOAuthApp: vi.fn(),
-      loginWithPersonalAccessToken: vi.fn(),
-      logoutFromAccount: vi.fn(),
-
-      fetchNotifications: vi.fn(),
-      removeAccountNotifications: vi.fn(),
-
-      markNotificationsAsRead: vi.fn(),
-      markNotificationsAsDone: vi.fn(),
-      unsubscribeNotification: vi.fn(),
-
-      isOnline: true,
-
-      shortcutRegistrationError: null,
-      clearShortcutRegistrationError: vi.fn(),
-
-      ...value,
-    } as TestAppContext;
-  }, [value]);
-
-  return (
-    <MemoryRouter initialEntries={initialEntries}>
-      <AppContext.Provider value={defaultValue as AppContextState}>{children}</AppContext.Provider>
-    </MemoryRouter>
-  );
-}
+const LOGIN_KEYS = [
+  'loginWithDeviceFlowStart',
+  'loginWithDeviceFlowPoll',
+  'loginWithDeviceFlowComplete',
+  'loginWithOAuthApp',
+  'loginWithPersonalAccessToken',
+  'logoutFromAccount',
+] as const;
 
 /**
  * Custom render that wraps components with all providers needed for testing:
- * MemoryRouter, QueryClientProvider, AppContext, and Zustand stores.
+ * MemoryRouter, QueryClientProvider, and Zustand stores.
  *
- * Account and settings state is seeded directly into the Zustand stores;
- * the remaining options override the (slim) AppContext values.
+ * Account, settings, and filter state is seeded directly into the Zustand
+ * stores; the remaining options override the globally mocked `useNotifications`,
+ * `useLogins`, and `useOnlineStatus` hooks (see hook-mocks.ts).
  *
  * Usage:
  *   renderWithProviders(<MyComponent />, { notifications, accounts, settings, filters, ... })
  */
 export function renderWithProviders(
   ui: ReactElement,
-  { initialEntries, accounts, settings, filters, ...context }: RenderOptions = {},
+  {
+    initialEntries,
+    accounts,
+    settings,
+    filters,
+    fetchNotifications,
+    isOnline,
+    shortcutRegistrationError,
+    ...hookOverrides
+  }: RenderOptions = {},
 ) {
   if (accounts) {
     useAccountsStore.setState({ accounts });
@@ -115,6 +81,32 @@ export function renderWithProviders(
   if (filters) {
     useFiltersStore.setState(filters);
   }
+
+  if (isOnline !== undefined) {
+    setIsOnlineOverride(isOnline);
+  }
+
+  if (shortcutRegistrationError !== undefined) {
+    useShortcutRegistrationStore.getState().setShortcutRegistrationError(shortcutRegistrationError);
+  }
+
+  const loginsOverrides: Partial<LoginsState> = {};
+  const notificationsOverrides: Partial<NotificationsState> = {};
+
+  for (const [key, value] of Object.entries(hookOverrides)) {
+    if ((LOGIN_KEYS as readonly string[]).includes(key)) {
+      loginsOverrides[key as keyof LoginsState] = value as never;
+    } else {
+      notificationsOverrides[key as keyof NotificationsState] = value as never;
+    }
+  }
+
+  if (fetchNotifications) {
+    notificationsOverrides.refetchNotifications = fetchNotifications;
+  }
+
+  setNotificationsOverrides(notificationsOverrides);
+  setLoginsOverrides(loginsOverrides);
 
   const queryClient = new QueryClient({
     defaultOptions: {
@@ -129,9 +121,7 @@ export function renderWithProviders(
   return render(ui, {
     wrapper: ({ children }) => (
       <QueryClientProvider client={queryClient}>
-        <AppContextProvider initialEntries={initialEntries} value={context}>
-          {children}
-        </AppContextProvider>
+        <MemoryRouter initialEntries={initialEntries}>{children}</MemoryRouter>
       </QueryClientProvider>
     ),
   });
