+++ src/renderer/components/notifications/RepositoryNotifications.tsx
@@ -3,7 +3,7 @@ import { type FC, type MouseEvent, useState } from 'react';
 import { CheckIcon, ReadIcon } from '@primer/octicons-react';
 import { Button, Stack } from '@primer/react';
 
-import { useAppContext } from '../../hooks/useAppContext';
+import { useNotifications } from '../../hooks/useNotifications';
 
 import { HoverButton } from '../primitives/HoverButton';
 import { HoverGroup } from '../primitives/HoverGroup';
@@ -27,7 +27,7 @@ export const RepositoryNotifications: FC<RepositoryNotificationsProps> = ({
   repoName,
   repoNotifications,
 }) => {
-  const { markNotificationsAsRead, markNotificationsAsDone } = useAppContext();
+  const { markNotificationsAsRead, markNotificationsAsDone } = useNotifications();
 
   const [shouldAnimateRepositoryExit, setShouldAnimateRepositoryExit] = useState(false);
   const [isRepositoryNotificationsVisible, setIsRepositoryNotificationsVisible] = useState(true);
