+++ src/renderer/components/primitives/Header.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import { ArrowLeftIcon, type Icon } from '@primer/octicons-react';
 import { IconButton, Stack } from '@primer/react';
 
-import { useAppContext } from '../../hooks/useAppContext';
+import { useNotifications } from '../../hooks/useNotifications';
 
 import { Title } from './Title';
 
@@ -17,7 +17,7 @@ interface HeaderProps {
 export const Header: FC<HeaderProps> = (props: HeaderProps) => {
   const navigate = useNavigate();
 
-  const { fetchNotifications } = useAppContext();
+  const { refetchNotifications: fetchNotifications } = useNotifications();
 
   return (
     <div className="pl-4 pr-5 pt-3 pb-1">
