--- src/renderer/hooks/useAppContext.ts
@@ -1,15 +1,0 @@
-import { useContext } from 'react';
-
-import { AppContext, type AppContextState } from '../context/context';
-
-/**
- * Custom hook that provides type-safe access to AppContext.
- * Throws if used outside of AppProvider.
- */
-export function useAppContext(): AppContextState {
-  const context = useContext(AppContext);
-  if (!context) {
-    throw new Error('useAppContext must be used within an AppProvider');
-  }
-  return context as AppContextState;
-}
